Moved disconnect test to SnippetListActivityTests
