+++ app/src/androidTest/java/com/microsoft/graph/snippets/SnippetListActivityTests.java
@@ -4,6 +4,8 @@
  */
 package com.microsoft.graph.snippets;
 
+import android.support.test.InstrumentationRegistry;
+import android.support.test.espresso.intent.Intents;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.widget.ListAdapter;
@@ -11,7 +13,9 @@ import android.widget.ListView;
 
 import com.microsoft.graph.snippets.snippet.AbstractSnippet;
 
+import org.junit.After;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -19,11 +23,31 @@ import org.junit.runner.RunWith;
 import java.util.ArrayList;
 import java.util.List;
 
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.Espresso.openActionBarOverflowOrOptionsMenu;
+import static android.support.test.espresso.action.ViewActions.click;
+import static android.support.test.espresso.intent.Intents.intended;
+import static android.support.test.espresso.intent.matcher.ComponentNameMatchers.hasShortClassName;
+import static android.support.test.espresso.intent.matcher.IntentMatchers.hasComponent;
+import static android.support.test.espresso.intent.matcher.IntentMatchers.toPackage;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static org.hamcrest.core.AllOf.allOf;
+
 @RunWith(AndroidJUnit4.class)
 public class SnippetListActivityTests {
     public static int CURRENT_NUMBER_OF_CATEGORIES = 6;
     @Rule
     public ActivityTestRule<SnippetListActivity> mSnippetListActivityRule = new ActivityTestRule<>(SnippetListActivity.class, false, false);
+
+    @Before
+    public void initIntents(){
+        Intents.init();
+    }
+
+    @After
+    public void releaseIntents() {
+        Intents.release();
+    }
 
     @Test
     public void checkRightNumberOfCategories() {
@@ -35,6 +59,28 @@ public class SnippetListActivityTests {
                 CURRENT_NUMBER_OF_CATEGORIES,
                 adapterSize - actualSnippets
         );
+    }
+
+    @Test
+    public void Disconnect() {
+        Disconnect(mSnippetListActivityRule);
+    }
+
+    private void Disconnect(ActivityTestRule<SnippetListActivity> snippetListActivityTestRule) {
+        SnippetListActivity snippetListActivity = snippetListActivityTestRule.launchActivity(null);
+
+        openActionBarOverflowOrOptionsMenu(InstrumentationRegistry.getTargetContext());
+
+        // Espresso can't find menu items by id. We'll use the text property.
+        onView(withText(R.string.disconnect_menu_item))
+                .perform(click());
+
+        intended(allOf(
+                hasComponent(hasShortClassName(".SignInActivity")),
+                toPackage("com.microsoft.graph.snippets")
+        ));
+
+        snippetListActivity.finish();
     }
 
     private int getItemsCount(ActivityTestRule<SnippetListActivity> snippetListActivityRule){
